fix output precisions

+++ examples/arkode/C_serial/ark_damped_harmonic_symplectic.c
@@ -136,8 +136,8 @@ int main(int argc, char* argv[])
   tret = T0;
   tout = T0 + dTout;
   /* Output current integration status */
-  fprintf(stdout, "t = %.6Lf, q(t) = %.6Lf, H = %.6Lf\n", tret, ydata[1],
-          Hamiltonian(y, tret));
+  fprintf(stdout, "t = %.6Lf, q(t) = %.6Lf, H = %.6Lf\n", (long double)tret,
+          (long double)ydata[1], (long double)Hamiltonian(y, tret));
 
   /* Do integration */
   for (iout = 0; iout < num_output_times; iout++)
@@ -146,8 +146,8 @@ int main(int argc, char* argv[])
     retval = SPRKStepEvolve(arkode_mem, tout, y, &tret, ARK_NORMAL);
 
     /* Output current integration status */
-    fprintf(stdout, "t = %.6Lf, q(t) = %.6Lf, H = %.6Lf\n", tret, ydata[1],
-            Hamiltonian(y, tret));
+    fprintf(stdout, "t = %.6Lf, q(t) = %.6Lf, H = %.6Lf\n", (long double)tret,
+            (long double)ydata[1], (long double)Hamiltonian(y, tret));
 
     /* Check if the solve was successful, if so, update the time and continue */
     if (retval >= 0)
